+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # GoFormation Versioning Changelog
 
+## [4.15.5](https://github.com/awslabs/goformation/compare/v4.15.4...v4.15.5) (2020-11-06)
+
+
+### Bug Fixes
+
+* **schema:** CloudFormation Updates ([#331](https://github.com/awslabs/goformation/issues/331)) ([12f9c83](https://github.com/awslabs/goformation/commit/12f9c835cbd9b980e06e30c458f95b14e16a3771))
+
+## [4.15.4](https://github.com/awslabs/goformation/compare/v4.15.3...v4.15.4) (2020-11-01)
+
+
+### Bug Fixes
+
+* **schema:** CloudFormation Updates ([#330](https://github.com/awslabs/goformation/issues/330)) ([4070319](https://github.com/awslabs/goformation/commit/40703191771425b3519128027478cf740d10f2d9))
+
+## [4.15.3](https://github.com/awslabs/goformation/compare/v4.15.2...v4.15.3) (2020-10-23)
+
+
+### Bug Fixes
+
+* **schema:** CloudFormation Updates ([#329](https://github.com/awslabs/goformation/issues/329)) ([4c1362b](https://github.com/awslabs/goformation/commit/4c1362bac14a3d6fc2cebd56eddcaf440228a4e9))
+
 ## [4.15.2](https://github.com/awslabs/goformation/compare/v4.15.1...v4.15.2) (2020-10-11)